Refactor: Move and rename /src/Airtable/u tils.js
Move the file `utils.js` from `/src/Airtable` to `/src/util` and rename it to `dates.js`
